+++ lib/gamefic/scriptable/queries.rb
@@ -12,7 +12,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::General]
       def anywhere *args, ambiguous: false
-        Query::General.new -> { entities }, *unproxy(args), ambiguous: ambiguous
+        Query::General.new -> { entities }, *args, ambiguous: ambiguous
       end
 
       # Define a query that searches an actor's family of entities. The
@@ -22,7 +22,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def available *args, ambiguous: false
-        Query::Scoped.new Scope::Family, *unproxy(args), ambiguous: ambiguous
+        Query::Scoped.new Scope::Family, *args, ambiguous: ambiguous
       end
       alias family available
 
@@ -31,7 +31,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def parent *args, ambiguous: false
-        Query::Scoped.new Scope::Parent, *unproxy(args), ambiguous: ambiguous
+        Query::Scoped.new Scope::Parent, *args, ambiguous: ambiguous
       end
 
       # Define a query that searches an actor's children.
@@ -39,7 +39,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def children *args, ambiguous: false
-        Query::Scoped.new Scope::Children, *unproxy(args), ambiguous: ambiguous
+        Query::Scoped.new Scope::Children, *args, ambiguous: ambiguous
       end
 
       # Define a query that searches an actor's siblings.
@@ -47,7 +47,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def siblings *args, ambiguous: false
-        Query::Scoped.new Scope::Siblings, *unproxy(args), ambiguous: ambiguous
+        Query::Scoped.new Scope::Siblings, *args, ambiguous: ambiguous
       end
 
       # Define a query that returns the actor itself.
@@ -55,7 +55,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def myself *args, ambiguous: false
-        Query::Scoped.new Scope::Myself, *unproxy(args), ambiguous: ambiguous
+        Query::Scoped.new Scope::Myself, *args, ambiguous: ambiguous
       end
 
       # Define a query that performs a plaintext search. It can take a String