+++ lib/gamefic/scriptable/plot_proxies.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Gamefic
+  module Scriptable
+    module PlotProxies
+      def plot_attr key
+        Proxy.new(:attr, [:plot, key])
+      end
+    end
+  end
+end