+++ lib/gamefic/subplot.rb
@@ -7,6 +7,8 @@ module Gamefic
   # started and concluded at any time during the parent plot's runtime.
   #
   class Subplot < Narrative
+    extend Scriptable::PlotProxies
+
     # @return [Hash]
     attr_reader :config