+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Configurable scanners
 - Refactored scanners and queries
 - Allow assignment to nil instance variables in stage
+- Lazy proxies
 
 ## 3.4.0 - September 10, 2024
 - Chapters